# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a PLAccounting API Server.",
        "title": "PLAccounting API",
        "contact": {},
        "version": "v1"
    },
    "paths": {
...
        "/book/{bid}/accountTitle/{tid}/transactions": {
            "get": {
                "description": "Get Sub Transactions from Account Title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sub Transaction"
                ],
                "summary": "Get Sub Transactions from Account Title",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account Title ID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sub Transactions was found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/book/{bid}/accountTitle/{tid}/transactions/{pid}": {
            "get": {
                "description": "Get Sub Transactions from Account Title with Page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sub Transaction"
                ],
                "summary": "Get Sub Transactions from Account Title with Page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Account Title ID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page ID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sub Transactions was found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/book/{bid}/transaction": {
            "get": {
                "description": "Get Transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get Transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transactions was found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Create Transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create Transaction",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/endpoint.CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created Transaction",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/book/{bid}/transaction/page/{pid}": {
            "get": {
                "description": "Get Transactions with Page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get Transactions with Page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page ID",
                        "name": "pid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transactions was found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/book/{bid}/transaction/{tid}": {
            "get": {
                "description": "Get Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get Transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction was found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Delete Transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction was deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Update Transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "bid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Transaction",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/endpoint.UpdateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated Transaction",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Request is failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "endpoint.CreateTransactionRequest": {
            "type": "object",
            "required": [
                "description",
                "occured_at",
                "sub_transactions"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "occured_at": {
                    "type": "string"
                },
                "sub_transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SubTransaction"
                    }
                }
            }
        },
...
        "endpoint.UpdateTransactionRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "occured_at": {
                    "type": "string"
                },
                "sub_transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SubTransaction"
                    }
                }
            }
        },
        "model.AccountTitle": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "amount_base": {
                    "type": "integer"
                },
                "book_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sub_transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SubTransaction"
                    }
                },
                "title_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.SubTransaction": {
            "type": "object",
            "properties": {
                "account_title": {
                    "$ref": "#/definitions/model.AccountTitle"
                },
                "account_title_id": {
                    "type": "integer"
                },
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "is_debit": {
                    "type": "boolean"
                },
                "sub_transaction_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}